DEV: restructure tests: move Helpers tests
