+++ tests/functional/helpers/CheckDatabaseJsonValuesTest.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace ls\tests;
+namespace ls\tests\functional\helpers;
+
+use ls\tests\TestBaseClass;
 
 /**
  * Check the JSON saved in database.
